+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **`[env]`** Added method `Env.Has`
 - **`[timeutil]`** Added method `TinyDate.IsZero`
 - **`[support/network]`** Added support of multiple resolvers
+- **`[hashutil]`** Added method `Hash.EqualConstantTime`
 - **`[color]`** Improved relative luminance calculation
 - **`[ansi]`** Removed `Codes` from all methods names
 - **`[ease]`** Fixed `SineIn` and `SineOut` calculation